requirements.txt
asgiref==3.8.1
Django==4.2.7
djangorestframework==3.12.3
sqlparse==0.5.3
tzdata==2024.2
django-cors-headers==4.6.0
drf-yasg==1.21.8
django-extensions==3.2.3
djangorestframework-simplejwt==5.3.1
pillow==11.0.0
django-filter
celery==5.2.3
django-haystack==3.0
kavenegar